+++ package.json
@@ -30,7 +30,7 @@
     "translate": "node scripts/translate.js",
     "push-translate": "tx push -s",
     "pull-translate": "tx pull -a -f && npm run translate",
-    "build-plugin": "export GENERATE_SOURCEMAP=false && npm run clean && npm run build && node scripts/build-plugin.js",
+    "build-plugin": "export GENERATE_SOURCEMAP=false && npm run clean && npm run format && npm run build && node scripts/build-plugin.js",
     "format": "prettier --write .",
     "type-check": "tsc"
   },
