+++ src/components/PluginPresets/PresetInput/index.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable react-hooks/exhaustive-deps */
 
 import React, { useEffect, useRef, useState } from 'react';
-import styles2 from '../../../styles/template-styles/Presets.module.scss';
+import styles2 from '../../../styles/template-styles/PluginPresets.module.scss';
 import { IPresetInput } from '../../../utils/Interfaces/PluginPresets/Input.interface';
 import { KeyDownActions } from '../../../utils/constants';
